Got a version that worked for the criterium racing league.
However, this doesn't accommodate multiple sprint/KOM segments that
are present in the various other leagues, so switching to using the
JSON results from ZwiftPower, as the CSV results don't list all of
the segments for sprint/KOMs. The parsing of the JSON is currently
working, now need to redo the league calculations.